refactor: move pkg-info type and module
Move pkg-info type into corresponding module. Also move that module into types folder.
