+++ src/cmd-search.ts
@@ -5,8 +5,8 @@ import log from "./logger";
 import { is404Error, isHttpError } from "./utils/error-type-guards";
 import * as os from "os";
 import assert from "assert";
-import { GlobalOptions, PkgInfo } from "./types/global";
-import { tryGetLatestVersion } from "./utils/pkg-info";
+import { GlobalOptions } from "./types/global";
+import { PkgInfo, tryGetLatestVersion } from "./types/pkg-info";
 import { env, parseEnv } from "./utils/env";
 import { DomainName } from "./types/domain-name";
 import { SemanticVersion } from "./types/semantic-version";
