+++ src/types/pkg-info.ts
@@ -1,5 +1,26 @@
-import { PkgInfo } from "../types/global";
-import { SemanticVersion } from "../types/semantic-version";
+import { SemanticVersion } from "./semantic-version";
+import { DomainName } from "./domain-name";
+import { PkgVersionInfo } from "./global";
+
+export type PkgInfo = {
+  name: DomainName;
+  _id?: DomainName;
+  _rev?: string;
+  _attachments?: Record<string, unknown>;
+  readme?: string;
+  versions: Record<SemanticVersion, PkgVersionInfo>;
+  "dist-tags"?: { latest?: SemanticVersion };
+  version?: SemanticVersion;
+  description?: string;
+  keywords?: string[];
+  time: {
+    [key: SemanticVersion]: string;
+    created?: string;
+    modified?: string;
+  };
+  date?: Date;
+  users?: Record<string, unknown>;
+};
 
 const hasLatestDistTag = (
   pkgInfo: Partial<PkgInfo>
