+++ test/mock-registry.ts
@@ -1,9 +1,9 @@
-import { PkgInfo } from "../src/types/global";
 import nock from "nock";
 import { SearchEndpointResult } from "./types";
 import { domainName, isDomainName } from "../src/types/domain-name";
 import assert from "assert";
 import { registryUrl } from "../src/types/registry-url";
+import { PkgInfo } from "../src/types/pkg-info";
 
 export const unityRegistryUrl = registryUrl("https://packages.unity.com");
 export const exampleRegistryUrl = registryUrl("http://example.com");
